--- Assets/Scripts/Tickets/TicketButtons.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 9cc939d008a813047975e8c67e2170f1
